+++ src/lib/declarative/form-fields/boolean/index.ts
@@ -1,0 +1,2 @@
+export { FieldCheckbox } from './field-checkbox'
+export { FieldSwitch } from './field-switch'