+++ src/lib/declarative/form-fields/specialized/index.ts
@@ -1,0 +1,7 @@
+export { FieldAddress } from './field-address'
+export { FieldCity } from './field-city'
+export { type ColorPickerFieldProps, FieldColorPicker } from './field-color-picker'
+export { FieldFileUpload, type FileUploadFieldProps } from './field-file-upload'
+export { FieldOTPInput } from './field-otp-input'
+export { FieldPhone } from './field-phone'
+export { FieldPinInput, type PinInputFieldProps } from './field-pin-input'